+++ parsonaut/lazy.py
